refactor: relanshionship in card and deck are now ower responsability

+++ src/main/java/br/com/cardgame/jeff/dtos/DeckDtoCreate.java
@@ -2,6 +2,6 @@ package br.com.cardgame.jeff.dtos;
 
 import java.time.LocalDateTime;
 
-public record DeckDtoCreate(LocalDateTime createIn, byte[] image64, String imgType,Long id ,String name){
+public record DeckDtoCreate(LocalDateTime createIn, byte[] image64, String imgType,int id ,String name){
     
 }
